# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Portfolio API",
        "description": "This is the Swagger UI documentation for my Laravel API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/information": {
            "get": {
                "tags": [
                    "Information"
                ],
                "summary": "Get all information",
                "operationId": "454b7f1fd588b5830ed076faa92816d2",
                "responses": {
                    "200": {
                        "description": "List of information"
                    }
                }
            },
            "post": {
                "tags": [
                    "Information"
                ],
                "summary": "Create a new information",
                "operationId": "3c90d1e04ca28d976112959f946a04b5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "i_img",
                                    "i_title",
                                    "i_detail",
                                    "display"
                                ],
                                "properties": {
                                    "i_img": {
                                        "type": "integer"
                                    },
                                    "i_title": {
                                        "type": "string"
                                    },
                                    "i_detail": {
                                        "type": "string"
                                    },
                                    "display": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Information created successfully"
                    }
                }
            }
        },
        "/api/information/{id}": {
            "get": {
                "tags": [
                    "Information"
                ],
                "summary": "Get an information by ID",
                "operationId": "17bb7a08fe95ca3122ffbc1e55e566ad",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Information retrieved successfully"
                    },
                    "404": {
                        "description": "Information not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Information"
                ],
                "summary": "Delete an information by ID",
                "operationId": "a6302c01324905a076e34df1c0c1aef7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "delete information successfully"
                    },
                    "404": {
                        "description": "Information not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Information"
                ],
                "summary": "Update a information by ID",
                "operationId": "65b91550c36aef633cfe070a8d865585",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "i_img": {
                                        "type": "integer"
                                    },
                                    "i_title": {
                                        "type": "string"
                                    },
                                    "i_detail": {
                                        "type": "string"
                                    },
                                    "display": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Information updated successfully"
                    },
                    "404": {
                        "description": "Information not found"
                    }
                }
            }
        },
...
    },
...
}